:truck: Move db from admin to dev
