--- lamindb/admin/__init__.py
@@ -1,8 +1,0 @@
-"""Admin control: Setup the database & update the schema.
-
-.. autosummary::
-   :toctree: .
-
-   db
-"""
-from . import db  # noqa
