+++ lamindb/dev/__init__.py
@@ -3,6 +3,7 @@
 .. autosummary::
    :toctree: .
 
+   db
    file
    object
 
@@ -16,5 +17,6 @@ Utilities:
 
 from lamindb_schema import id  # noqa
 
+from . import db  # noqa
 from . import file, object  # noqa
 from ._docs import doc_args  # noqa
